Ajustes de queries lentas, colocando no contexto de leitura
