+++ src/SME.SGP.Dados/Repositorios/RepositorioFeriadoCalendario.cs
@@ -4,6 +4,7 @@ using SME.SGP.Dominio;
 using SME.SGP.Dominio.Interfaces;
 using SME.SGP.Infra;
 using SME.SGP.Infra.Interface;
+using SME.SGP.Infra.Interfaces;
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
@@ -12,7 +13,7 @@ namespace SME.SGP.Dados.Repositorios
 {
     public class RepositorioFeriadoCalendario : RepositorioBase<FeriadoCalendario>, IRepositorioFeriadoCalendario
     {
-        public RepositorioFeriadoCalendario(ISgpContext conexao, IServicoAuditoria servicoAuditoria) : base(conexao, servicoAuditoria)
+        public RepositorioFeriadoCalendario(ISgpContextConsultas conexao, IServicoAuditoria servicoAuditoria) : base(conexao, servicoAuditoria)
         {
         }
