+++ src/SME.SGP.Dominio.Interfaces/Repositorios/IRepositorioFechamentoTurma.cs
@@ -1,7 +1,4 @@
-﻿using SME.SGP.Infra.Dtos;
-using System.Threading.Tasks;
-
-namespace SME.SGP.Dominio.Interfaces
+﻿namespace SME.SGP.Dominio.Interfaces
 {
     public interface IRepositorioFechamentoTurma : IRepositorioBase<FechamentoTurma>
     {
